server/requirements.txt
fastapi[standard]
pydantic
langchain
langchain-google-genai
pymysql
sqlalchemy
deepgram-sdk
python-dotenv
uagents==0.22.3
uagents-adapter==0.4.1